+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix generation of JWT signature.
+
 ### Added
 
 - new method on JWT `unknownKid()` to generate a random value for the `kid`