+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- When signing a JWT, a specific kid can be specified. This allows to create
+  tokens which are signed with an incorrect key
+
 ## [0.10.0] - 2024-06-17
 
 ### Fixed